chore(lint): fix linting issues
