+++ src/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Game from './containers/game';
-import registerServiceWorker from './registerServiceWorker';
 import reportWebVitals from './reportWebVitals';
 import './index.scss';
 
@@ -12,9 +11,6 @@ ReactDOM.render(
     </React.StrictMode>, 
     document.getElementById('root')
 );
-
-/* istanbul ignore next */
-registerServiceWorker();
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
